fix(composition): Scroll by wheel deltas finer than a detent
InteractionTracker took whole wheel detents as an int, divided from the
120-unit MouseWheelDelta with integer division, so the small deltas a
precision touchpad reports floored to zero and never scrolled a
ScrollView. Carry the detents as a double down the whole chain.

+++ src/Uno.UI.Composition/Composition/InteractionTracker/InteractionTracker.cs
@@ -153,7 +153,8 @@ public partial class InteractionTracker : CompositionObject
 		_state.ReceiveInertiaStarting(-linearVelocity);
 	}
 
-	internal void ReceivePointerWheel(int mouseWheelTicks, bool isHorizontal)
+	/// <param name="mouseWheelTicks">Detents scrolled, fractional for touchpads that report deltas finer than one detent.</param>
+	internal void ReceivePointerWheel(double mouseWheelTicks, bool isHorizontal)
 	{
 		// On WinUI, this depends on mouse setting "how many lines to scroll each time"
 		// The default Windows setting is 3 lines, and each line is 16px.
